requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
python-dotenv==0.15.0
WTForms==2.3.3
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
Flask-Bootstrap==3.3.7.1
boto3==1.16.29
Flask-Login==0.5.0
stripe==2.55.1
termcolor==1.1.0
plaid-python==7.2.0
psycopg2-binary==2.8.6
